Ensure native data is large enough for a pointer on 64-bit machines and new object layout.

+++ lib/jamvm/java/lang/VMClassLoaderData.java
@@ -23,7 +23,7 @@ import java.util.ArrayList;
 
 public class VMClassLoaderData {
     ArrayList unloaders;
-    int hashtable;
+    long hashtable;
 
     synchronized void newLibraryUnloader(long dllEntry) {
         if(unloaders == null)
